Finite le eccezioni

+++ MODEL/ciclismo.h
@@ -14,8 +14,8 @@ private:
 public:
     Ciclismo(unsigned int =0, unsigned int = 0, unsigned int = 0, unsigned int = 0, unsigned int = 0);
     Ciclismo(const Ciclismo&);
-    virtual double calorie() const;
-    virtual double GrassiBruc() const;
+    virtual unsigned int calorie() const;
+    virtual unsigned int GrassiBruc() const;
     unsigned int get_distanza() const;
     unsigned int get_pianura() const;
     unsigned int get_salita() const;
